+++ examples/esmodule_events/mod.ts
@@ -9,7 +9,7 @@ declare global {
 }
 
 const _hyperEvents = new HyperEvents({
-	target: document,
+	host: document,
 	connected: true,
 	eventNames: ["click", "pointerover"],
 });
